Bump phpunit and PSR standards
