+++ .travis.yml
@@ -2,13 +2,6 @@ language: php
 
 matrix:
   include:
-    - php: 5.5.9
-      dist: trusty
-      env: 'COMPOSER_FLAGS="--prefer-stable --prefer-lowest"'
-    - php: 5.5
-      dist: trusty
-    - php: 5.6
-      dist: xenial
     - php: 7.0
       dist: xenial
     - php: 7.1
@@ -24,5 +17,5 @@ install:
   - travis_retry composer update ${COMPOSER_FLAGS}
 
 script:
-  - vendor/bin/phpcs --standard=psr2 src/
+  - vendor/bin/phpcs --standard=psr12 src/
   - vendor/bin/phpunit
