+++ composer.json
@@ -12,22 +12,22 @@
         }
     ],
     "require": {
-        "php": ">=5.5.9",
+        "php": "^7.0",
         "symfony/http-foundation": "~2.7|~3.0|~4.0|~5.0",
         "symfony/http-kernel": "~2.7|~3.0|~4.0|~5.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.0 || ^4.8.10",
+        "phpunit/phpunit": "^6|^7|^8|^9",
         "squizlabs/php_codesniffer": "^3.5"
     },
     "autoload": {
         "psr-4": {
-            "Asm89\\Stack\\": "src/Asm89/Stack/"
+            "Asm89\\Stack\\": "src/"
         }
     },
     "autoload-dev": {
         "psr-4": {
-            "Asm89\\Stack\\": "test/Asm89/Stack/"
+            "Asm89\\Stack\\Tests\\": "tests/"
         }
     },
     "scripts": {
